Pequeño error de enunciado que hace que no carguen las texturas bien

+++ Practica/IG1App/Mesh.h
@@ -46,7 +46,7 @@ protected:
 
 	//EDITABLE
 
-	std::vector<glm::dvec2> vTexCoords;
+	std::vector<glm::vec2> vTexCoords;
 
 	virtual void draw() const;
